chore: rename to file-viewer
