+++ src/components/file-viewer/file-viewer.scss
